Fixed compiler warnings about comparison issues between 'int' and unsigned int (aka 'size_t')
In other words converted many loops in the form:

    for(int i = 0; i< schoolBuildings->size();i++){

to:

    for (size_t i = 0; i < schoolBuildings->size(); i++) {

The changes are:

1. Convert 'int' to 'size_t' which is the data type used by all C++
   data structures, because negative indexes or sizes (suggested by
   int) is not meangiful and only unsigned numbers should be
   used. Sadly, many languages do not support unsigned numbers and
   that forces programmers to have bad mental model and develop
   tererible code.

2. Spacing -- in English there should be 1 blank space between words.
   Similarly in C++ there should be a blank space between very
   word/operator as shown above.  Of course, these types of changes
   are automatically done by the editor (emacs/Eclipse/NetBeans etc.)
   but the programmer has to click 2 buttons for that happen.

   I have not reformatted the code fully to keep the deltas small to
   avoid breaking the code.

+++ haplos/haplos/src/ImageGeneration/XFigImageGenerator.cpp
@@ -60,8 +60,8 @@ void XFigImageGenerator::createImage(const std::string& outFileName,
 	const int maxColors = generateColorScale(xfig);
 	const int radius = 4, scale = 4;
 	// Iterate over the points and generate circles...
-	for (size_t r = 0; (r < rows); r++) {
-		for (size_t c = 0; (c < cols); c++) {
+	for (int r = 0; (r < rows); r++) {
+		for (int c = 0; (c < cols); c++) {
 			// Compute color code for
 			if (data[r][c] <= 0) {
 				if(data[r][c]==0){
